+++ src/main/java/duke/command/InvalidCommand.java
@@ -1,7 +1,7 @@
 package duke.command;
 
-import duke.DukeException;
-import duke.Storage;
+import duke.exception.DukeException;
+import duke.storage.Storage;
 import duke.Ui;
 import duke.task.TaskList;
