+++ src/main/java/duke/command/ToDoCommand.java
@@ -2,7 +2,7 @@ package duke.command;
 
 import java.io.IOException;
 
-import duke.Storage;
+import duke.storage.Storage;
 import duke.Ui;
 import duke.task.Task;
 import duke.task.TaskList;
